requirements.txt
attrs==21.4.0
lark==0.12.0
pathlib2==2.3.5
SQLAlchemy==1.4.54
alembic==1.5.5
enum==0.4.7
enum34==1.1.10
pendulum==2.1.2
voluptuous==0.13.1
repoze.lru==0.7
PyYAML==5.4.1
icecream==2.1.3
structlog==20.1.0
better-exceptions==0.3.3
Faker==3.0.1
flask==1.1.4
